requirements.txt
pillow==11.3.0
python-dotenv==1.2.1
requests==2.32.5
